+++ pypeit/tests/test_spectrographs.py
@@ -11,7 +11,7 @@ from pypeit.pypmsgs import PypeItError
 from pypeit import spectrographs
 from pypeit.spectrographs.util import load_spectrograph
 from pypeit import pypeitsetup
-from pypeit.par.util import make_pypeit_file
+from pypeit.tests import tstutils
 from pypeit.tests.tstutils import data_path
 
 
@@ -28,9 +28,10 @@ def test_shanekastblue():
 
 
 def test_select_detectors_pypeit_file():
-    # Generate a PYPIT file
+    # Generate a PypeIt file
+    pypeItFile = tstutils.make_shane_kast_blue_pypeitfile()
     pypeit_file = data_path('test.pypeit')
-    make_pypeit_file(pypeit_file, 'shane_kast_blue', [data_path('b*fits.gz')], setup_mode=True)
+    pypeItFile.write(pypeit_file)
 
     # Perform the setup
     setup = pypeitsetup.PypeItSetup.from_pypeit_file(pypeit_file)